+++ sdwan_feature_templates.tf
@@ -154,6 +154,25 @@ resource "sdwan_cedge_global_feature_template" "cedge_global_feature_template" {
   depends_on                    = [sdwan_localized_policy.localized_policy]
 }
 
+resource "sdwan_cedge_igmp_feature_template" "cedge_igmp_feature_template" {
+  for_each     = { for t in try(local.edge_feature_templates.igmp_templates, {}) : t.name => t }
+  name         = each.value.name
+  description  = each.value.description
+  device_types = [for d in try(each.value.device_types, local.defaults.sdwan.edge_feature_templates.igmp_templates.device_types) : try(local.device_type_map[d], "vedge-${d}")]
+  interfaces = try(length(each.value.interfaces) == 0, true) ? null : [for interface in each.value.interfaces : {
+    name          = try(interface.name, null)
+    name_variable = try(interface.name_variable, null)
+    optional      = try(interface.optional, null)
+    join_groups = !can(interface.join_groups) ? null : [for group in interface.join_groups : {
+      group_address          = try(group.group_address, null)
+      group_address_variable = try(group.group_address_variable, null)
+      source                 = try(group.source, null)
+      source_variable        = try(group.source_variable, null)
+    }]
+  }]
+  depends_on = [sdwan_localized_policy.localized_policy]
+}
+
 resource "sdwan_cisco_banner_feature_template" "cisco_banner_feature_template" {
   for_each       = { for t in try(local.edge_feature_templates.banner_templates, {}) : t.name => t }
   name           = each.value.name