+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - add `expanded_community_list_variable` parameter support in `sdwan_route_policy_definition`
 - add support for sdwan_policy_object_app_probe_class resource
 - update UX 2.0 enum values to be lowercase wherever possible
+- add support for sdwan_cedge_igmp_feature_template resource
 - add support for sdwan_cedge_multicast_feature_template resource
 
 ## 1.1.0